requirements.txt
Django==3.1
djangorestframework==3.12
psycopg2
django-registration
